fixed wrong position of conversation window

+++ Mac/Controller.h
@@ -51,7 +51,7 @@
 - (void)openNewMessageWindowInReplyTo:(NSString *)userName statusId:(NSString *)statusId withString:(NSString *)string;
 - (NSString *)pluginURL;
 - (void)handleGetURLEvent:(NSAppleEventDescriptor *)event withReplyEvent:(NSAppleEventDescriptor *)replyEvent;
-- (void)unreadMentions:(NSInteger)count;
+- (void)unreadMentions:(int)count;
 - (void)notificateUserAboutMention:(NSString *)text fromName:(NSString *)name withPostId:(NSString *)postId andEntity:(NSString *)entity;
 
 - (void)openURL:(NSString *)url;
