+++ Mac/Controller.m
@@ -247,7 +247,7 @@
 	return nil;
 }
 
-- (void)unreadMentions:(NSInteger)count {
+- (void)unreadMentions:(int)count {
 	if (![mentionsViewWindow isVisible] && count > 0) {
 		[timelineViewWindow setTitle:[NSString stringWithFormat:@"Tentia (^%i)", count]];
 		[[[NSApplication sharedApplication] dockTile] setBadgeLabel:[NSString stringWithFormat:@"%i", count]];
